requirements.txt
chainlit
langchain
langchain-community
langchain-text-splitters
langchain-ollama
faiss-cpu
tree-sitter>=0.20.4,<0.22.0
tree-sitter-languages
PyYAML
unstructured[pdf,docx,md]
pypdf
esprima